Fit a Latent Semantic Scaling model

Latent Semantic Scaling (LSS) is a word embedding-based semisupervised algorithm for document scaling.

Usage

textmodel_lss(x, ...)

# S3 method for dfm
textmodel_lss(
  x,
  seeds,
  terms = NULL,
  k = 300,
  slice = NULL,
  weight = "count",
  cache = FALSE,
  simil_method = "cosine",
  engine = c("RSpectra", "irlba", "rsvd"),
  auto_weight = FALSE,
  include_data = FALSE,
  group_data = FALSE,
  verbose = FALSE,
  ...
)

# S3 method for fcm
textmodel_lss(
  x,
  seeds,
  terms = NULL,
  w = 50,
  max_count = 10,
  weight = "count",
  cache = FALSE,
  simil_method = "cosine",
  engine = c("rsparse"),
  auto_weight = FALSE,
  verbose = FALSE,
  ...
)

Arguments

x

a dfm or fcm created by quanteda::dfm() or quanteda::fcm()

...

additional arguments passed to the underlying engine.

seeds

a character vector or named numeric vector that contains seed words. If seed words contain "*", they are interpreted as glob patterns. See quanteda::valuetype.

terms

a character vector or named numeric vector that specify words for which polarity scores will be computed; if a numeric vector, words' polarity scores will be weighted accordingly; if NULL, all the features of quanteda::dfm() or quanteda::fcm() will be used.

k

the number of singular values requested to the SVD engine. Only used when x is a dfm.

slice

a number or indices of the components of word vectors used to compute similarity; slice < k to further truncate word vectors; useful for diagnosys and simulation.

weight

weighting scheme passed to quanteda::dfm_weight(). Ignored when engine is "rsparse".

cache

if TRUE, save result of SVD for next execution with identical x and settings. Use the base::options(lss_cache_dir) to change the location cache files to be save.

simil_method

specifies method to compute similarity between features. The value is passed to quanteda.textstats::textstat_simil(), "cosine" is used otherwise.

engine

select the engine to factorize x to generate word vectors. Choose from RSpectra::svds(), irlba::irlba(), rsvd::rsvd(), and rsparse::GloVe().

auto_weight

automatically determine weights to approximate the polarity of terms to seed words. See details.

include_data

if TRUE, fitted model includes the dfm supplied as x.

group_data

if TRUE, apply dfm_group(x) before saving the dfm.

verbose

show messages if TRUE.

w

the size of word vectors. Used only when x is a fcm.

max_count

passed to x_max in rsparse::GloVe$new() where cooccurrence counts are ceiled to this threshold. It should be changed according to the size of the corpus. Used only when x is a fcm.

Details

Latent Semantic Scaling (LSS) is a semisupervised document scaling method. textmodel_lss() constructs word vectors from use-provided documents (x) and weights words (terms) based on their semantic proximity to seed words (seeds). Seed words are any known polarity words (e.g. sentiment words) that users should manually choose. The required number of seed words are usually 5 to 10 for each end of the scale.

If seeds is a named numeric vector with positive and negative values, a bipolar LSS model is construct; if seeds is a character vector, a unipolar LSS model. Usually bipolar models perform better in document scaling because both ends of the scale are defined by the user.

A seed word's polarity score computed by textmodel_lss() tends to diverge from its original score given by the user because it's score is affected not only by its original score but also by the original scores of all other seed words. If auto_weight = TRUE, the original scores are weighted automatically using stats::optim() to minimize the squared difference between seed words' computed and original scores. Weighted scores are saved in seed_weighted in the object.

References

Watanabe, Kohei. 2020. "Latent Semantic Scaling: A Semisupervised Text Analysis Technique for New Domains and Languages", Communication Methods and Measures. doi:10.1080/19312458.2020.1832976 .

Watanabe, Kohei. 2017. "Measuring News Bias: Russia's Official News Agency ITAR-TASS' Coverage of the Ukraine Crisis" European Journal of Communication. doi:10.1177/0267323117695735 .

Examples

# \donttest{
library("quanteda")
#> Package version: 3.2.5
#> Unicode version: 13.0
#> ICU version: 66.1
#> Parallel computing: 4 of 4 threads used.
#> See https://quanteda.io for tutorials and examples.
library("LSX")

# download corpus
corp <- tryCatch({
   con <- url("https://bit.ly/2GZwLcN", "rb")
   readRDS(con)
   },
   error = function(e) e,
   warning = function(w) w,
   finally = close(con)
)

if (!exists("corp"))
   quit("no")

toks <- corpus_reshape(corp, "sentences") %>%
        tokens(remove_punct = TRUE) %>%
        tokens_remove(stopwords("en")) %>%
        tokens_select("^[\\p{L}]+$", valuetype = "regex", padding = TRUE)
dfmt <- dfm(toks) %>%
        dfm_trim(min_termfreq = 10)

seed <- as.seedwords(data_dictionary_sentiment)

# SVD
lss_svd <- textmodel_lss(dfmt, seed)
head(coef(lss_svd), 20)
#>   superior   flirting  excellent visitation dedication       good     humour 
#> 0.13155587 0.12173367 0.11775988 0.11280701 0.10336445 0.10096614 0.09880631 
#>     hinges   menswear       nice      highs bernadette    gilford     julian 
#> 0.09686758 0.09634406 0.09474779 0.09357691 0.09341598 0.09227394 0.09113390 
#>  specialty      tepid   observer injunction   positive   listener 
#> 0.09087273 0.09082654 0.09002514 0.08903475 0.08826408 0.08797926 
head(predict(lss_svd, newdata = dfmt))
#> text120761.1 text120761.2 text120761.3 text120761.4 text120761.5 text120761.6 
#>    0.9583329    0.1300070    1.6720468    0.5208033   -0.9270076    0.1614789 
head(predict(lss_svd, newdata = dfmt, min_n = 10)) # more robust
#> text120761.1 text120761.2 text120761.3 text120761.4 text120761.5 text120761.6 
#>    1.3038419    0.1393459    2.3072114    0.6887440   -0.6426615    0.1835903 

dfmt_grp <- dfm_group(dfmt) # group sentences

# sentiment on economy
eco <- head(textstat_context(toks, 'econom*'), 500)
lss_svd_eco <- textmodel_lss(dfmt, seed, terms = eco)
head(predict(lss_svd_eco, newdata = dfmt_grp))
#> text120761 text174574 text141269 text151432 text169265 text134827 
#>  1.7540268  2.3001292  0.3976127  0.2266591 -0.3749561  0.3671907 

# sentiment on politics
pol <- head(textstat_context(toks, 'politi*'), 500)
lss_svd_pol <- textmodel_lss(dfmt, seed, terms = pol)
head(predict(lss_svd_pol, newdata = dfmt_grp))
#>  text120761  text174574  text141269  text151432  text169265  text134827 
#> -0.48447935  1.79613965  0.09528141 -1.07422468 -0.10760545 -0.83658239 

# modify hyper-parameters of existing model
lss_svd_pol2 <- as.textmodel_lss(lss_svd_pol, seed[c(1, 8)], terms = pol, slice = 200)
head(predict(lss_svd_pol2, newdata = dfmt_grp))
#>  text120761  text174574  text141269  text151432  text169265  text134827 
#>  0.89925290 -0.81232153 -2.19564710  0.11160452 -1.69063750 -0.02687483 

# GloVe
fcmt <- fcm(dfmt, tri = FALSE)
lss_glov <- textmodel_lss(fcmt, seed)
#> INFO  [21:34:10.016] epoch 1, loss 0.0822
#> INFO  [21:34:17.044] epoch 2, loss 0.0634
#> INFO  [21:34:23.932] epoch 3, loss 0.0574
#> INFO  [21:34:30.674] epoch 4, loss 0.0544
#> INFO  [21:34:39.164] epoch 5, loss 0.0526
#> INFO  [21:34:46.384] epoch 6, loss 0.0514
#> INFO  [21:34:53.273] epoch 7, loss 0.0505
#> INFO  [21:35:00.221] epoch 8, loss 0.0499
#> INFO  [21:35:07.243] epoch 9, loss 0.0493
#> INFO  [21:35:14.182] epoch 10, loss 0.0489
head(predict(lss_glov, newdata = dfmt_grp))
#> text120761 text174574 text141269 text151432 text169265 text134827 
#>  0.3933700  0.8595003  0.2900752  0.6007621  0.6505581 -1.8600892 
# }